Move DI and Middleware config to Program.cs Remove Discord.NET dependency in services project Upgrade all projects to .NET 8
